Got OpenGL to render CUDA Image

+++ Path Tracer CUDA/Texture.h
@@ -1,14 +1,13 @@
 #pragma once
 #include <iostream>
 #include <glad/glad.h>
-
 using namespace std;
 
 class Texture {
 public:
-	Texture(const uint8_t* data, uint32_t width, uint32_t height);
+	Texture(const float* data, uint32_t width, uint32_t height);
 	void bind(unsigned int slot = 0);
-	void load(const uint8_t* data);
+	void load(const float* data);
 	void unbind();
 	~Texture();
 private:
